+++ src/configuration.cpp
@@ -132,6 +132,7 @@ void Configuration::Load()
     this->PerfNetworkUseBits =  s.value("Performance/NetworkUseBits",    this->PerfNetworkUseBits).toBool();
     this->PerfGraphWindowSec =  s.value("Performance/GraphWindowSec",    this->PerfGraphWindowSec).toInt();
     this->PerfSidePanelGroupOrder = s.value("Performance/SidePanelGroupOrder", this->PerfSidePanelGroupOrder).toStringList();
+    this->PerformanceSplitterState = s.value("Performance/SplitterState",     this->PerformanceSplitterState).toByteArray();
 
     // For now this is hardcoded, we may want to make it customizable later
     this->RefreshRateAvailableIntervals = { 250, 500, 1000, 2000, 5000, 15000 };
@@ -203,6 +204,7 @@ void Configuration::Save()
     s.setValue("Performance/NetworkUseBits",            this->PerfNetworkUseBits);
     s.setValue("Performance/GraphWindowSec",            this->PerfGraphWindowSec);
     s.setValue("Performance/SidePanelGroupOrder",       this->PerfSidePanelGroupOrder);
+    s.setValue("Performance/SplitterState",             this->PerformanceSplitterState);
 
     s.sync();
 }